.ci/scripts/generate_cc_changelog/requirements.txt
jinja2
pygithub
gitpython